# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/config": {
            "get": {
                "description": "Get app config",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Get App Config",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/generate-report": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Generate report from sent results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "projectId",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "executionName",
                        "name": "execution_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "executionName",
                        "name": "execution_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "executionName",
                        "name": "execution_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "description": "Get All projects",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get All projects",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/projects/{project_id}": {
            "get": {
                "description": "Get Project By ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get Project By ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "default",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Create Project",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Create Project",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "default",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete Project",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Delete Project",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "default",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/projects/{project_id}/reports/{path}": {
            "get": {
                "description": "Get Path in a project directory",
                "consumes": [
                    "*/*"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get Path in a project directory",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "default",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "latest/widgets/summary.json",
                        "description": "default",
                        "name": "path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/send-results": {
            "post": {
                "description": "Send allure result files to server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Send results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "projectId",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "create project if not exists",
                        "name": "force_project_creation",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "result files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}